reduce a lot from einsum logics
